# apps/web/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Strategy Terminal",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/watchlist/entries": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "List Watchlist Entries",
        "operationId": "listWatchlistEntries",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryListResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Create Watchlist Entry",
        "operationId": "createWatchlistEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistEntryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Watchlist entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Watchlist conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Watchlist input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/watchlist/entries/order": {
      "put": {
        "tags": [
          "watchlist"
        ],
        "summary": "Reorder Watchlist Entries",
        "operationId": "reorderWatchlistEntries",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistEntriesReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Watchlist entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Watchlist conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Watchlist input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/watchlist/entries/{entry_id}": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Get Watchlist Entry",
        "operationId": "getWatchlistEntry",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Watchlist entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Watchlist conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Watchlist input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "watchlist"
        ],
        "summary": "Patch Watchlist Entry",
        "operationId": "patchWatchlistEntry",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Entry Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistEntryPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Watchlist entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Watchlist conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Watchlist input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "watchlist"
        ],
        "summary": "Delete Watchlist Entry",
        "operationId": "deleteWatchlistEntry",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Entry Id"
            }
          },
          {
            "name": "expected_revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Watchlist entry not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Watchlist conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Watchlist input rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Watchlist service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetType": {
        "type": "string",
        "enum": [
          "stock",
          "etf",
          "index",
          "sector_index",
          "fund",
          "futures",
          "fx",
          "unknown"
        ],
        "title": "AssetType"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "validation": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Validation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "Exchange": {
        "type": "string",
        "enum": [
          "SSE",
          "SZSE",
          "BSE",
          "HKEX",
          "NASDAQ",
          "NYSE",
          "AMEX",
          "OTHER"
        ],
        "title": "Exchange"
      },
...
      "InstrumentId": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Symbol"
          },
          "exchange": {
            "$ref": "#/components/schemas/Exchange"
          },
          "market": {
            "$ref": "#/components/schemas/Market"
          },
          "asset_type": {
            "$ref": "#/components/schemas/AssetType"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "default": "CNY"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbol",
          "exchange",
          "market",
          "asset_type"
        ],
        "title": "InstrumentId"
      },
...
      "Market": {
        "type": "string",
        "enum": [
          "CN",
          "HK",
          "US",
          "GLOBAL"
        ],
        "title": "Market"
      },
...
      "ValidationIssue": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "issue_type"
        ],
        "title": "ValidationIssue"
      },
      "WatchlistEntriesReorderRequest": {
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "securities",
              "sectors"
            ],
            "title": "Scope"
          },
          "role": {
            "$ref": "#/components/schemas/WatchlistRole"
          },
          "ordered_items": {
            "items": {
              "$ref": "#/components/schemas/WatchlistOrderItemRequest"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Ordered Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scope",
          "role",
          "ordered_items"
        ],
        "title": "WatchlistEntriesReorderRequest"
      },
      "WatchlistEntryCreateRequest": {
        "properties": {
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "role": {
            "$ref": "#/components/schemas/WatchlistRole"
          },
          "display_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name"
          },
          "note": {
            "type": "string",
            "maxLength": 256,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "instrument",
          "role",
          "display_name"
        ],
        "title": "WatchlistEntryCreateRequest",
        "description": "Closed configuration request; it performs no market-data side effect."
      },
      "WatchlistEntryListResponse": {
        "properties": {
          "configuration_only": {
            "type": "boolean",
            "const": true,
            "title": "Configuration Only",
            "default": true
          },
          "monitoring_active": {
            "type": "boolean",
            "const": false,
            "title": "Monitoring Active",
            "default": false
          },
          "alert_rule_created": {
            "type": "boolean",
            "const": false,
            "title": "Alert Rule Created",
            "default": false
          },
          "market_data_requested": {
            "type": "boolean",
            "const": false,
            "title": "Market Data Requested",
            "default": false
          },
          "trade_created": {
            "type": "boolean",
            "const": false,
            "title": "Trade Created",
            "default": false
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "schema_version": {
            "type": "string",
            "const": "watchlist-list/v1",
            "title": "Schema Version",
            "default": "watchlist-list/v1"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WatchlistEntryView"
            },
            "type": "array",
            "title": "Items"
          },
          "limits": {
            "$ref": "#/components/schemas/WatchlistLimitsView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "WatchlistEntryListResponse"
      },
      "WatchlistEntryPatchRequest": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WatchlistRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "WatchlistEntryPatchRequest"
      },
      "WatchlistEntryResponse": {
        "properties": {
          "configuration_only": {
            "type": "boolean",
            "const": true,
            "title": "Configuration Only",
            "default": true
          },
          "monitoring_active": {
            "type": "boolean",
            "const": false,
            "title": "Monitoring Active",
            "default": false
          },
          "alert_rule_created": {
            "type": "boolean",
            "const": false,
            "title": "Alert Rule Created",
            "default": false
          },
          "market_data_requested": {
            "type": "boolean",
            "const": false,
            "title": "Market Data Requested",
            "default": false
          },
          "trade_created": {
            "type": "boolean",
            "const": false,
            "title": "Trade Created",
            "default": false
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "schema_version": {
            "type": "string",
            "const": "watchlist-entry/v1",
            "title": "Schema Version",
            "default": "watchlist-entry/v1"
          },
          "action": {
            "type": "string",
            "enum": [
              "created",
              "read",
              "updated",
              "deleted"
            ],
            "title": "Action"
          },
          "entry": {
            "$ref": "#/components/schemas/WatchlistEntryView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action",
          "entry"
        ],
        "title": "WatchlistEntryResponse"
      },
      "WatchlistEntryView": {
        "properties": {
          "entry_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Entry Id"
          },
          "instrument": {
            "$ref": "#/components/schemas/InstrumentId"
          },
          "role": {
            "$ref": "#/components/schemas/WatchlistRole"
          },
          "display_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name"
          },
          "note": {
            "type": "string",
            "maxLength": 256,
            "title": "Note"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entry_id",
          "instrument",
          "role",
          "display_name",
          "note",
          "created_at",
          "updated_at",
          "revision"
        ],
        "title": "WatchlistEntryView"
      },
      "WatchlistLimitsView": {
        "properties": {
          "max_core_stocks": {
            "type": "integer",
            "const": 10,
            "title": "Max Core Stocks",
            "default": 10
          },
          "max_active_stocks": {
            "type": "integer",
            "const": 30,
            "title": "Max Active Stocks",
            "default": 30
          },
          "max_active_entries": {
            "type": "integer",
            "const": 100,
            "title": "Max Active Entries",
            "default": 100
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WatchlistLimitsView"
      },
      "WatchlistOrderItemRequest": {
        "properties": {
          "entry_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Entry Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "entry_id",
          "expected_revision"
        ],
        "title": "WatchlistOrderItemRequest"
      },
      "WatchlistRole": {
        "type": "string",
        "enum": [
          "core",
          "watchlist",
          "etf",
          "index"
        ],
        "title": "WatchlistRole"
      },
...
    }
  }
}